get apikey REST API user authentication working
